.ai/requirements.txt
mcp
pyyaml
pathlib